backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
pydantic[email]==2.6.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
supabase==2.3.7
httpx==0.26.0
loguru==0.7.2
pytest==8.0.0
pytest-asyncio==0.23.5
APScheduler==3.10.4
